% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mspec.R
\name{plot.SpectralClMixed}
\alias{plot.SpectralClMixed}
\title{Plots the output of mspec}
\usage{
\method{plot}{SpectralClMixed}(x,cols=NULL,...)
}
\arguments{
\item{x}{object of SpectralClMixed class, the output of mspec}
\item{cols}{For datasets with more than 10 columns, columns to plot}
\item{...}{other graphic parameters}
}
\value{No return value,the function produces a parallel coordinate plot and a scatter plot matrix}
\description{
Plots the output of the function mspec, which performs Spectral clustering for mixed type data. The function displays up to 10 variables on a parrallel coordinate plot and on a scatter plot matrix, with colors representing the clustering partition
}
\examples{
ex1=mspec(iris,3)
 plot(ex1,cols=1:4)}