% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MixSpeC.R
\docType{package}
\name{SpectralClMix}
\alias{SpectralClMix}
\title{SpectralClMix}
\description{
Cluster analysis of mixed-type data using Spectral Clustering.
}
\references{F. Mbuga and, C. Tortora. Spectral Clustering of Mixed-Type Data. \emph{Stats}, 5(1) 2022  }
\author{
Felix Mbuga, Cristina Tortora, Zander Bonnet
}
