% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mspec.R
\name{summary.SpectralClMixed}
\alias{summary.SpectralClMixed}
\title{Summarizes the output of mspec}
\usage{
\method{summary}{SpectralClMixed}(object,...)
}
\arguments{
\item{object}{object of SpectralClMixed class, the output of mspec}
\item{...}{other optional  parameters}
}
\value{It displays: The selected categorical variables weight, The between divided by within sum of squares,  The total within sum of squares, and the cluster size.
}
\description{
Summarizes the output of mspec
}
